adressing issue rte-france#126 and refactoring Runner module
