+++ grid2op/Runner/__init__.py
@@ -2,5 +2,5 @@ __all__ = [
     "Runner"
 ]
 
-from grid2op.Runner.Runner import Runner
+from grid2op.Runner.runner import Runner
